+++ SMS - Copy/model/Product.cs
@@ -6,9 +6,9 @@ namespace SMS.model
         public int Id { get; set; }
         public string BarCode { get; set; }
         public string ProductName { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int ProductQuantity{get;set;}
-        public Product(int id, string barCode, string productName, double price, int productQuantity)
+        public Product(int id, string barCode, string productName, decimal price, int productQuantity)
         {
             Id = id;
             BarCode = barCode;
