+++ SMS - Copy/model/zCustomer.cs
@@ -6,14 +6,14 @@ namespace SMS.model
         public int Id { get; set; }
         public string Address { get; set; }
         public string CustomerId { get; set; }
-        public double Wallet { get; set; }
+        public decimal Wallet { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
         public int Pin { get; set; }
-        // public Customer(int id, string firstName, string lastName, int staffId, string email, string phoneNumber, int pin, string post, string address, string dateOfBirth, int age, double wallet) : base(id, firstName, lastName, staffId, email, phoneNumber, pin, dateOfBirth, age)
-        public Customer(int id, string firstName, string lastName, string email, string customerId, string phoneNumber, int pin, string address, double wallet)
+        // public Customer(int id, string firstName, string lastName, int staffId, string email, string phoneNumber, int pin, string post, string address, string dateOfBirth, int age, decimal wallet) : base(id, firstName, lastName, staffId, email, phoneNumber, pin, dateOfBirth, age)
+        public Customer(int id, string firstName, string lastName, string email, string customerId, string phoneNumber, int pin, string address, decimal wallet)
         {
             Id = id;
             FirstName = firstName;
