+++ SMS/interfaces/ITransactionManager.cs
@@ -2,13 +2,13 @@ namespace SMS.interfaces
 {
     public interface ITransactionManager
     {
-        void CreateTransaction(string barCode, int quantity, string customerId, double cashTender);
+        void CreateTransaction(string barCode, int quantity, string customerId, decimal cashTender);
         // void GetTransaction(string receiptNo);
         // Transaction UpdateTransaction();
         // void DeleteTransaction(string receiptNo);
         void GetAllTransactions();
-        double GetAllTransactionsAdmin();
-        double CalculateTotalSales();
+        decimal GetAllTransactionsAdmin();
+        decimal CalculateTotalSales();
         // void ReadFromFile();
         // void ReWriteToFile();
     }
