+++ SMS/interfaces/zICustomerManager.cs
@@ -5,7 +5,7 @@ namespace SMS.interfaces
 {
     public interface ICustomerManager
     {
-        void CreateCustomer(string firstName, string lastName, string email, string phoneNumber, int pin, string address,double wallet);
+        void CreateCustomer(string firstName, string lastName, string email, string phoneNumber, int pin, string address,decimal wallet);
         Customer GetCustomer(string customerId);
         void UpdateCustomer(string customerId, string firstName, string lastName, string phoneNumber);
         void DeleteCustomer(string customerId);
