+++ SMS/model/Product.cs
@@ -5,9 +5,9 @@ namespace SMS.model
         // public int Id { get; set; }
         public string BarCode { get; set; }
         public string ProductName { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int ProductQuantity { get; set; }
-        public Product( string barCode, string productName, double price, int productQuantity)
+        public Product( string barCode, string productName, decimal price, int productQuantity)
         {
             // Id = id;
             BarCode = barCode;
@@ -15,15 +15,5 @@ namespace SMS.model
             Price = price;
             ProductQuantity = productQuantity;
         }
-        // public string WriteToFIle()
-        // {
-        //     return $"{Id}%%%%%{BarCode}%%%%%{ProductName}%%%%%{Price}%%%%%{ProductQuantity}";
-        // }
-
-        // public static Product ConvertToProduct(string productAllFromText)
-        // {
-        //     var productConvert = productAllFromText.Split("%%%%%");
-        //     return new Product(int.Parse(productConvert[0]), productConvert[1], productConvert[2], int.Parse(productConvert[3]), int.Parse(productConvert[4]));
-        // }
     }
 }
