+++ SMS/model/Wallet.cs
@@ -3,11 +3,11 @@ namespace SMS.model
     public class Wallet
     {
         public int Id { get; set; }
-        public double WalletRemainingBalance { get; set; }
-        public double WalletTotal { get; set; }
-        public double WalletWithdrawal { get; set; }
+        public decimal WalletRemainingBalance { get; set; }
+        public decimal WalletTotal { get; set; }
+        public decimal WalletWithdrawal { get; set; }
 
-        public Wallet(int id, double walletTotal, double walletWithdrawal, double walletRemainingBalance)
+        public Wallet(int id, decimal walletTotal, decimal walletWithdrawal, decimal walletRemainingBalance)
         {
             Id = id;
             WalletTotal = walletTotal;
